pyproject.toml
[project]
name = "pythonfastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "peewee>=3.18.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
